% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{cdfg_port_samples}
\alias{cdfg_port_samples}
\title{Chinook genotype data from 1974 individuals sampled by CA Dept of Fish and Game at California Ports}
\format{This is a data frame with 1974 rows and 190 columns.  Each row contains
the genotype of a Chinook salmon that was caught in the ocean and sampled at a port.
It was used for some analyses in Clemento et al.  The row name is the name given to the
fish.  The 190 columns are the alleles carried at each of the 95 loci.  This is in two-column
format, so every locus gets two adjacent columns.  The column headers have the names of the loci.  The
first occurrence of each locus is just the locus name and the second has a ".1" appended to it.

In general, alleles are coded as follows: A=1, C=2, G=3, T=4.  There are some loci that get an allele
code of 5, which refers to some data feature of some sort.  You can think of it as an allele.
There are only two alleles observed at each locus.  Missing data are denoted by NAs.}
\usage{
data(cdfg_port_samples)
}
\description{
Chinook genotype data from 1974 individuals sampled by CA Dept of Fish and Game at California Ports
}
\references{
Clemento AJ, Crandall ED, Garza JC, Anderson EC (2014) Evaluation of a single nucleotide polymorphism baseline for genetic stock identification of Chinook Salmon (Oncorhynchus tshawytscha) in the California Current Large Marine Ecosystem. Fishery Bulletin 112(2-3): 112-130. doi:10.7755/FB.112.2-3.2
}
\keyword{datasets}

