% Generated by roxygen2 (4.0.1): do not edit by hand
\name{P_likelihood}
\alias{P_likelihood}
\title{Compute likelihoods for populations of origin}
\usage{
P_likelihood(gc, genos, lambda)
}
\arguments{
\item{snp_zeroes}{Matrix of counts of the "zero" allele at L SNPs in P populations.
This is an L x P integer matrix where L is the number of SNP
loci and P is the number of populations in the baseline. he observed number of 0 alleles
at locus j in population p is snp_zeroes[j,p]}

\item{snp_ones}{Same as above, but for the 1 allele at each SNP.}

\item{genos}{Fill in.}

\item{lambda}{Fill in.}
}
\description{
More later
}

