% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{SNPcontam}
\alias{SNPcontam}
\alias{SNPcontam-package}
\title{Detecting contaminated samples from SNP genotypes}
\description{
\code{SNPcontam} is a package in development.
It really is just in development at the moment
}

